Устранено предупреждение C4018 после r5895.
git-svn-id: svn://svn.code.sf.net/p/mpcbe/code/trunk@5903 7426f7ed-705a-47fc-8cda-d59144bb0dce

+++ src/apps/mplayerc/AddCommandDlg.cpp
@@ -57,7 +57,7 @@ BOOL CAddCommandDlg::OnInitDialog()
 
 	CAppSettings& s = AfxGetAppSettings();
 
-	for (int i = 0; i < s.wmcmds.size(); i++) {
+	for (int i = 0; i < (int)s.wmcmds.size(); i++) {
 		const wmcmd& wc = s.wmcmds[i];
 
 		m_list.InsertItem(i, wc.GetName());
